add method calculateCalories

+++ app/src/main/java/com/malytic/altituden/fragments/GraphFragment.java
@@ -47,7 +47,7 @@ public class GraphFragment extends Fragment {
         graph.removeAllSeries();
         if(MainActivity.pathData.elevation != null && MainActivity.pathData.elevation.size() > 0) {
             int dataSetSize = MainActivity.pathData.elevation.size();
-            float dx = MainActivity.pathData.length / dataSetSize;
+            float dx = (float)MainActivity.pathData.length / (float)dataSetSize;
             DataPoint[] dataPoints = new DataPoint[dataSetSize];
             for (int i = 0; i < dataPoints.length; i++) {
                 dataPoints[i] = new DataPoint((dx * i), MainActivity.pathData.elevation.get(i).getElevation());
